Set serial as id noticegroup

+++ sql_schema/rid.sql
@@ -82,7 +82,7 @@ create table notice(
 );
 
 create table noticegroup(
-    id integer NOT NULL,
+    id bigserial NOT NULL,
     operator_id varchar(255),
     noticeRef integer references notice(id),
     primary key(id,noticeref)
